reformatting files using spring java code formatter

+++ src/main/java/com/spring5/entity/Product.java
@@ -13,7 +13,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import java.math.BigDecimal;
 import java.util.Map;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -22,8 +21,7 @@ import lombok.NoArgsConstructor;
 
 /**
  * @author david
- * @version $LastChangedRevision $LastChangedDate Last Modified Author:
- * $LastChangedBy
+ * @version $LastChangedRevision $LastChangedDate Last Modified Author: $LastChangedBy
  */
 @Data
 @Builder(toBuilder = true)
@@ -33,57 +31,54 @@ import lombok.NoArgsConstructor;
 @Table(name = "PRODUCT")
 public class Product {
 
-    public Product(long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
+	public Product(long id, String name) {
+		this.id = id;
+		this.name = name;
+	}
 
-    public Product(
-            long id, String name, BigDecimal price, int quantity, String description, boolean status) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.quantity = quantity;
-        this.description = description;
-        this.status = status;
-    }
+	public Product(long id, String name, Double price, int quantity, String description, boolean status) {
+		this.id = id;
+		this.name = name;
+		this.price = price;
+		this.quantity = quantity;
+		this.description = description;
+		this.status = status;
+	}
 
-    public Product(
-            Long id,
-            String name,
-            String category,
-            String description,
-            double basePrice,
-            Map<String, String[]> variants) {
-        this.id = id;
-        this.name = name;
-        this.category = category;
-        this.description = description;
-        this.basePrice = basePrice;
-        this.variants = variants;
-    }
+	public Product(Long id, String name, String category, String description, double basePrice,
+			Map<String, String[]> variants) {
+		this.id = id;
+		this.name = name;
+		this.category = category;
+		this.description = description;
+		this.basePrice = basePrice;
+		this.variants = variants;
+	}
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "ID")
-    private long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "ID")
+	private long id;
 
-    @Column
-    private String name;
+	@Column
+	private String name;
 
-    @Column
-    private BigDecimal price;
+	@Column
+	private Double price;
 
-    @Column
-    private int quantity;
+	@Column
+	private int quantity;
 
-    @Column
-    private String description;
+	@Column
+	private String description;
 
-    @Column
-    private boolean status;
+	@Column
+	private boolean status;
 
-    private String category;
-    private double basePrice;
-    private Map<String, String[]> variants;
+	private String category;
+
+	private double basePrice;
+
+	private Map<String, String[]> variants;
+
 }
